refactor(ExpenseDto): make Ids use longs

+++ src/main/java/ml/meajudadev/api/v1/dto/ExpenseDto.java
@@ -6,14 +6,13 @@ import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record ExpenseDto(
-        @JsonProperty("id") int id,
-        @JsonProperty("userId") int userId,
-        @JsonProperty("categoryId") Integer categoryId,
+        @JsonProperty("id") long id,
+        @JsonProperty("userId") long userId,
+        @JsonProperty("categoryId") long categoryId,
         @JsonProperty("description") String description,
         @JsonProperty("value") BigDecimal value,
         @JsonProperty("isFixed") boolean isFixed,
         @JsonProperty("dueDate") LocalDate dueDate,
         @JsonProperty("expenseType") char expenseType,
         @JsonProperty("expenseState") char expenseState
-) {
-}
+) { }
